+++ packages/vm/src/ffi.rs
@@ -7,7 +7,7 @@ use thiserror::Error;
 /// non-negligible computational cost and must always have gas information
 /// attached. In order to prevent new calls from forgetting such gas information
 /// to be passed, the inner success and failure types contain gas information.
-pub type FfiResult<T> = (core::result::Result<T, FfiError>, GasInfo);
+pub type BackendResult<T> = (core::result::Result<T, BackendError>, GasInfo);
 
 #[derive(Copy, Clone, Debug)]
 pub struct GasInfo {
@@ -54,62 +54,62 @@ impl AddAssign for GasInfo {
 }
 
 #[derive(Error, Debug)]
-pub enum FfiError {
+pub enum BackendError {
     #[error("Panic in FFI call")]
     ForeignPanic {},
-    #[error("bad argument passed to FFI")]
+    #[error("Bad argument")]
     BadArgument {},
     #[error("VM received invalid UTF-8 data from backend")]
     InvalidUtf8 {},
     #[error("Iterator with ID {id} does not exist")]
     IteratorDoesNotExist { id: u32 },
-    #[error("Ran out of gas during FFI call")]
+    #[error("Ran out of gas during call into backend")]
     OutOfGas {},
-    #[error("Unknown error during FFI call: {msg:?}")]
+    #[error("Unknown error during call into backend: {msg:?}")]
     Unknown { msg: Option<String> },
-    // This is the only error case of FfiError that is reported back to the contract.
-    #[error("User error during FFI call: {msg}")]
+    // This is the only error case of BackendError that is reported back to the contract.
+    #[error("User error during call into backend: {msg}")]
     UserErr { msg: String },
 }
 
-impl FfiError {
+impl BackendError {
     pub fn foreign_panic() -> Self {
-        FfiError::ForeignPanic {}
+        BackendError::ForeignPanic {}
     }
 
     pub fn bad_argument() -> Self {
-        FfiError::BadArgument {}
+        BackendError::BadArgument {}
     }
 
     pub fn iterator_does_not_exist(iterator_id: u32) -> Self {
-        FfiError::IteratorDoesNotExist { id: iterator_id }
+        BackendError::IteratorDoesNotExist { id: iterator_id }
     }
 
     pub fn out_of_gas() -> Self {
-        FfiError::OutOfGas {}
+        BackendError::OutOfGas {}
     }
 
     pub fn unknown<S: ToString>(msg: S) -> Self {
-        FfiError::Unknown {
+        BackendError::Unknown {
             msg: Some(msg.to_string()),
         }
     }
 
     /// Use `::unknown(msg: S)` if possible
     pub fn unknown_without_message() -> Self {
-        FfiError::Unknown { msg: None }
+        BackendError::Unknown { msg: None }
     }
 
     pub fn user_err<S: ToString>(msg: S) -> Self {
-        FfiError::UserErr {
+        BackendError::UserErr {
             msg: msg.to_string(),
         }
     }
 }
 
-impl From<FromUtf8Error> for FfiError {
+impl From<FromUtf8Error> for BackendError {
     fn from(_original: FromUtf8Error) -> Self {
-        FfiError::InvalidUtf8 {}
+        BackendError::InvalidUtf8 {}
     }
 }
 
@@ -142,63 +142,63 @@ mod test {
 
     #[test]
     fn ffi_error_foreign_panic() {
-        let error = FfiError::foreign_panic();
+        let error = BackendError::foreign_panic();
         match error {
-            FfiError::ForeignPanic { .. } => {}
+            BackendError::ForeignPanic { .. } => {}
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn ffi_error_bad_argument() {
-        let error = FfiError::bad_argument();
+        let error = BackendError::bad_argument();
         match error {
-            FfiError::BadArgument { .. } => {}
+            BackendError::BadArgument { .. } => {}
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn iterator_does_not_exist_works() {
-        let error = FfiError::iterator_does_not_exist(15);
+        let error = BackendError::iterator_does_not_exist(15);
         match error {
-            FfiError::IteratorDoesNotExist { id, .. } => assert_eq!(id, 15),
+            BackendError::IteratorDoesNotExist { id, .. } => assert_eq!(id, 15),
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn ffi_error_out_of_gas() {
-        let error = FfiError::out_of_gas();
+        let error = BackendError::out_of_gas();
         match error {
-            FfiError::OutOfGas { .. } => {}
+            BackendError::OutOfGas { .. } => {}
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn ffi_error_unknown() {
-        let error = FfiError::unknown("broken");
+        let error = BackendError::unknown("broken");
         match error {
-            FfiError::Unknown { msg, .. } => assert_eq!(msg.unwrap(), "broken"),
+            BackendError::Unknown { msg, .. } => assert_eq!(msg.unwrap(), "broken"),
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn ffi_error_unknown_without_message() {
-        let error = FfiError::unknown_without_message();
+        let error = BackendError::unknown_without_message();
         match error {
-            FfiError::Unknown { msg, .. } => assert!(msg.is_none()),
+            BackendError::Unknown { msg, .. } => assert!(msg.is_none()),
             e => panic!("Unexpected error: {:?}", e),
         }
     }
 
     #[test]
     fn ffi_error_user_err() {
-        let error = FfiError::user_err("invalid input");
+        let error = BackendError::user_err("invalid input");
         match error {
-            FfiError::UserErr { msg, .. } => assert_eq!(msg, "invalid input"),
+            BackendError::UserErr { msg, .. } => assert_eq!(msg, "invalid input"),
             e => panic!("Unexpected error: {:?}", e),
         }
     }
@@ -207,9 +207,9 @@ mod test {
 
     #[test]
     fn convert_from_fromutf8error() {
-        let error: FfiError = String::from_utf8(vec![0x80]).unwrap_err().into();
+        let error: BackendError = String::from_utf8(vec![0x80]).unwrap_err().into();
         match error {
-            FfiError::InvalidUtf8 { .. } => {}
+            BackendError::InvalidUtf8 { .. } => {}
             e => panic!("Unexpected error: {:?}", e),
         }
     }