+++ packages/vm/src/imports.rs
@@ -15,7 +15,7 @@ use crate::context::{
 };
 use crate::conversion::to_u32;
 use crate::errors::{CommunicationError, VmError, VmResult};
-use crate::ffi::FfiError;
+use crate::ffi::BackendError;
 #[cfg(feature = "iterator")]
 use crate::memory::maybe_read_region;
 use crate::memory::{read_region, write_region};
@@ -114,7 +114,9 @@ pub fn do_canonicalize_address<A: Api, S: Storage, Q: Querier>(
             write_region(ctx, destination_ptr, canonical.as_slice())?;
             Ok(0)
         }
-        Err(FfiError::UserErr { msg, .. }) => Ok(write_to_contract::<S, Q>(ctx, msg.as_bytes())?),
+        Err(BackendError::UserErr { msg, .. }) => {
+            Ok(write_to_contract::<S, Q>(ctx, msg.as_bytes())?)
+        }
         Err(err) => Err(VmError::from(err)),
     }
 }
@@ -134,7 +136,9 @@ pub fn do_humanize_address<A: Api, S: Storage, Q: Querier>(
             write_region(ctx, destination_ptr, human.as_str().as_bytes())?;
             Ok(0)
         }
-        Err(FfiError::UserErr { msg, .. }) => Ok(write_to_contract::<S, Q>(ctx, msg.as_bytes())?),
+        Err(BackendError::UserErr { msg, .. }) => {
+            Ok(write_to_contract::<S, Q>(ctx, msg.as_bytes())?)
+        }
         Err(err) => Err(VmError::from(err)),
     }
 }
@@ -230,7 +234,7 @@ mod test {
     };
     use crate::testing::{MockApi, MockQuerier, MockStorage};
     use crate::traits::Storage;
-    use crate::FfiError;
+    use crate::BackendError;
 
     static CONTRACT: &[u8] = include_bytes!("../testdata/contract.wasm");
 
@@ -623,8 +627,8 @@ mod test {
         let api = MockApi::new_failing("Temporarily unavailable");
         let result = do_canonicalize_address::<MA, MS, MQ>(api, ctx, source_ptr, dest_ptr);
         match result.unwrap_err() {
-            VmError::FfiErr {
-                source: FfiError::Unknown { msg, .. },
+            VmError::BackendErr {
+                source: BackendError::Unknown { msg, .. },
             } => {
                 assert_eq!(msg.unwrap(), "Temporarily unavailable");
             }
@@ -728,8 +732,8 @@ mod test {
         let api = MockApi::new_failing("Temporarily unavailable");
         let result = do_humanize_address::<MA, MS, MQ>(api, ctx, source_ptr, dest_ptr);
         match result.unwrap_err() {
-            VmError::FfiErr {
-                source: FfiError::Unknown { msg, .. },
+            VmError::BackendErr {
+                source: BackendError::Unknown { msg, .. },
             } => assert_eq!(msg.unwrap(), "Temporarily unavailable"),
             err => panic!("Incorrect error returned: {:?}", err),
         };
@@ -1028,8 +1032,8 @@ mod test {
         let non_existent_id = 42u32;
         let result = do_next::<MS, MQ>(ctx, non_existent_id);
         match result.unwrap_err() {
-            VmError::FfiErr {
-                source: FfiError::IteratorDoesNotExist { id, .. },
+            VmError::BackendErr {
+                source: BackendError::IteratorDoesNotExist { id, .. },
             } => assert_eq!(id, non_existent_id),
             e => panic!("Unexpected error: {:?}", e),
         }