+++ packages/vm/src/traits.rs
@@ -2,7 +2,7 @@ use cosmwasm_std::{Binary, CanonicalAddr, ContractResult, HumanAddr, SystemResul
 #[cfg(feature = "iterator")]
 use cosmwasm_std::{Order, KV};
 
-use crate::ffi::FfiResult;
+use crate::ffi::BackendResult;
 
 /// Holds all external dependencies of the contract.
 /// Designed to allow easy dependency injection at runtime.
@@ -22,7 +22,7 @@ pub trait Storage {
     ///
     /// Note: Support for differentiating between a non-existent key and a key with empty value
     /// is not great yet and might not be possible in all backends. But we're trying to get there.
-    fn get(&self, key: &[u8]) -> FfiResult<Option<Vec<u8>>>;
+    fn get(&self, key: &[u8]) -> BackendResult<Option<Vec<u8>>>;
 
     /// Allows iteration over a set of key/value pairs, either forwards or backwards.
     /// Returns an interator ID that is unique within the Storage instance.
@@ -36,24 +36,29 @@ pub trait Storage {
     /// The implementation must ensure that iterator IDs are assigned in a deterministic manner as this is
     /// environment data that is injected into the contract.
     #[cfg(feature = "iterator")]
-    fn scan(&mut self, start: Option<&[u8]>, end: Option<&[u8]>, order: Order) -> FfiResult<u32>;
+    fn scan(
+        &mut self,
+        start: Option<&[u8]>,
+        end: Option<&[u8]>,
+        order: Order,
+    ) -> BackendResult<u32>;
 
     /// Returns the next element of the iterator with the given ID.
     ///
-    /// If the ID is not found, a FfiError::IteratorDoesNotExist is returned.
+    /// If the ID is not found, a BackendError::IteratorDoesNotExist is returned.
     ///
     /// This call must not change data in the storage, but incrementing an iterator can be a mutating operation on
     /// the Storage implementation.
     #[cfg(feature = "iterator")]
-    fn next(&mut self, iterator_id: u32) -> FfiResult<Option<KV>>;
+    fn next(&mut self, iterator_id: u32) -> BackendResult<Option<KV>>;
 
-    fn set(&mut self, key: &[u8], value: &[u8]) -> FfiResult<()>;
+    fn set(&mut self, key: &[u8], value: &[u8]) -> BackendResult<()>;
 
     /// Removes a database entry at `key`.
     ///
     /// The current interface does not allow to differentiate between a key that existed
     /// before and one that didn't exist. See https://github.com/CosmWasm/cosmwasm/issues/290
-    fn remove(&mut self, key: &[u8]) -> FfiResult<()>;
+    fn remove(&mut self, key: &[u8]) -> BackendResult<()>;
 }
 
 /// Api are callbacks to system functions defined outside of the wasm modules.
@@ -66,8 +71,8 @@ pub trait Storage {
 /// We can use feature flags to opt-in to non-essential methods
 /// for backwards compatibility in systems that don't have them all.
 pub trait Api: Copy + Clone + Send {
-    fn canonical_address(&self, human: &HumanAddr) -> FfiResult<CanonicalAddr>;
-    fn human_address(&self, canonical: &CanonicalAddr) -> FfiResult<HumanAddr>;
+    fn canonical_address(&self, human: &HumanAddr) -> BackendResult<CanonicalAddr>;
+    fn human_address(&self, canonical: &CanonicalAddr) -> BackendResult<HumanAddr>;
 }
 
 pub trait Querier {
@@ -78,11 +83,11 @@ pub trait Querier {
     ///
     /// The gas limit describes how much VM gas this particular query is allowed
     /// to comsume when measured separately from the rest of the contract.
-    /// The returned gas info (in FfiResult) can exceed the gas limit in cases
+    /// The returned gas info (in BackendResult) can exceed the gas limit in cases
     /// where the query could not be aborted exactly at the limit.
     fn query_raw(
         &self,
         request: &[u8],
         gas_limit: u64,
-    ) -> FfiResult<SystemResult<ContractResult<Binary>>>;
+    ) -> BackendResult<SystemResult<ContractResult<Binary>>>;
 }