+++ CHANGELOG.md
@@ -58,6 +58,8 @@
 - Rename `Extern` to `Backend`.
 - Rename `mock_dependencies` to `mock_backend` and
   `mock_dependencies_with_balances` to `mock_backend_with_balances`.
+- Rename `FfiError`/`FfiResult` to `BackendError`/`BackendResult` and adapt
+  `VmError` accordingly.
 
 ## 0.11.2 (2020-10-26)